requirements.txt
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-phonenumber-field==2.2.0
gunicorn==19.9.0
idna==2.8
phonenumbers==8.10.7
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
